Safer decimal number (in)equality testing operators

The %d==%, %d!=% %d<%, %d>%, %d<=%, %d>=% (in)equality operator perform decimal (class "double") number truth testing.
They are virtually equivalent to the regular (in)equality operators,
==, !=, <, >, <=, >=,
except for one aspect:
The decimal number (in)equality operators assume that if the absolute difference between any two numbers x and y is smaller than the Machine tolerance, sqrt(.Machine$double.eps), then x and y should be consider to be equal.

Thus these operators provide safer decimal number (in)equality tests.

For example: 0.1*7 == 0.7 returns FALSE, even though they are equal, due to the way decimal numbers are stored in programming languages (like R). But 0.1*7 %d==% 0.7 returns TRUE.

There are also the x %d{}% bnd and x %d!{}% bnd operators, where bnd is a vector of length 2, or a 2-column matrix (nrow(bnd)==length(x) or nrow(bnd)==1).
The x %d{}% bnd operator checks if x is within the closed interval with bounds defined by bnd.
The x %d!{}% bnd operator checks if x is outside the closed interval with bounds defined by bnd.

x %d==% y

x %d!=% y

x %d<% y

x %d>% y

x %d<=% y

x %d>=% y

x %d{}% bnd

x %d!{}% bnd

tinyoperations_decimal_truth()

Arguments

x, y

numeric vectors, matrices, or arrays, though these operators were specifically designed for decimal numbers (class "double").

bnd

either a vector of length 2, or a matrix with 2 columns and 1 row, or else a matrix with 2 columns where nrow(bnd)==length(x).
The first element/column of bnd gives the lower bound of the closed interval;
The second element/column of bnd gives the upper bound of the closed interval;

Value

Same as ==.

See also

Examples

x <- c(0.3, 0.6, 0.7)
y <- c(0.1*3, 0.1*6, 0.1*7)
print(x); print(y)
#> [1] 0.3 0.6 0.7
#> [1] 0.3 0.6 0.7
x == y # gives FALSE, but should be TRUE
#> [1] FALSE FALSE FALSE
x!= y # gives TRUE, should be FALSE
#> [1] TRUE TRUE TRUE
x > y # not wrong
#> [1] FALSE FALSE FALSE
x < y # gives TRUE, should be FALSE
#> [1] TRUE TRUE TRUE
x %d==% y # here it's done correctly
#> [1] TRUE TRUE TRUE
x %d!=% y # correct
#> [1] FALSE FALSE FALSE
x %d<% y # correct
#> [1] FALSE FALSE FALSE
x %d>% y # correct
#> [1] FALSE FALSE FALSE
x %d<=% y # correct
#> [1] TRUE TRUE TRUE
x %d>=% y # correct
#> [1] TRUE TRUE TRUE

x <- c(0.3, 0.6, 0.7)
bnd <- matrix(c(0.29, 0.59, 0.69, 0.31, 0.61, 0.71), ncol=2)
x %d{}% bnd
#> [1] TRUE TRUE TRUE
x %d!{}% bnd
#> [1] FALSE FALSE FALSE

# These operators still work for non-decimal number numerics also:
x <- 1:5
y <- 1:5
x %d==% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d!=% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d<% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d>% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d<=% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d>=% y
#> [1] TRUE TRUE TRUE TRUE TRUE

x <- 1:5
y <- x+1
x %d==% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d!=% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d<% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d>% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d<=% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d>=% y
#> [1] FALSE FALSE FALSE FALSE FALSE

x <- 1:5
y <- x-1
x %d==% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d!=% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d<% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d>% y
#> [1] TRUE TRUE TRUE TRUE TRUE
x %d<=% y
#> [1] FALSE FALSE FALSE FALSE FALSE
x %d>=% y
#> [1] TRUE TRUE TRUE TRUE TRUE