(un)expose infix operators in the current environment

import_inops(expose=...) exposes infix operators specified in a package or an alias object to the current environment (like the global environment or the environment within a function).

import_inops(unexpose=...) unexposes (i.e. removes) the infix operators specified in a packages or an alias object from the current environment (like the global environment or the environment within a function).
Note that in this case only infix operators exposed by import_inops() will be removed from the current environment;
"regular" infix operators (i.e. user-specified) will not be touched.

import_inops(expose = NULL, unexpose = NULL, lib.loc = .libPaths(), ...)

Arguments

expose

either one of the following:

a package alias as produced by the import_as function.

a string giving the package name from which to load infix operators, and place them in the current environment.

unexpose

either one of the following:

a package alias as produced by the import_as function.

a string giving the package name.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
Only used when supplying a character vector of package names to expose / unexpose, and ignored when supplying an alias object to expose / unexpose (the library is already defined inside the alias object).
The lib.loc argument would usually be .libPaths().
See also loadNamespace.

...

additional arguments, only relevant if the expose argument is used.
See import_inops.control.

Value

If using argument expose:

The infix operators specified in the given package or alias will be placed in the current environment (like the Global environment, or the environment within a function).

If using argument unexpose:

The infix operators specified in the given package or alias, exposed by import_inops(), will be deleted.

If such infix operators could not be found, this function returns NULL.

Details

The import_inops() function does not support overloading base/core R operators, so don't try.

When using import_inops() to remove infix operators from the current environment, it will use the attributes of those operators to determine if the infix operator came from import_inops(), or if they were user-defined.

See also

Examples

if (FALSE) {
import_as(dt., "data.table")
import_inops(expose = dt.) # expose infix operators from alias
import_inops(unexpose = dt.) # unexposed infix operators from current environment
import_inops(expose = "data.table") # expose infix operators from package
import_inops(unexpose = "data.table") # remove the exposed infix operators from environment

# additional arguments (only used when exposing, not unexposing):
import_as(dt., "data.table")
import_inops(expose = dt., include.only = ":=")
import_inops(unexpose = dt.)
import_inops(expose = "data.table", exclude = ":=")
import_inops(unexpose = "data.table")

import_inops(expose = dt., overwrite = FALSE)
import_inops(unexpose = dt.)
import_inops(expose = "data.table", overwrite = FALSE)
import_inops(unexpose = "data.table")
}