import_inops.control

Additional arguments to control exposing infix operators in the import_inops function.

import_inops.control(
  exclude = NULL,
  include.only = NULL,
  overwrite = TRUE,
  inherits = FALSE
)

Arguments

exclude

a character vector, giving the infix operators NOT to expose to the current environment.
This can be handy to prevent overwriting any (user defined) infix operators already present in the current environment.

include.only

a character vector, giving the infix operators to expose to the current environment, and the rest of the operators will not be exposed.
This can be handy to prevent overwriting any (user defined) infix operators already present in the current environment.

overwrite

logical, indicating if it is allowed to overwrite existing infix operators.

If TRUE (default), a warning is given when operators existing in the current environment are being overwritten, but the function continuous nonetheless.

If FALSE, an error is produced when the to be exposed operators already exist in the current environment, and the function is halted.

inherits

logical; when overwrite=FALSE, should enclosed environments, especially package namespaces, also be taken into account?
Defaults to FALSE.
See also exists.

Value

This function is used internally in the import_inops function.

Details

You cannot specify both the exclude and include.only arguments. Only one or the other, or neither.

See also