In place modifying string subsetting

In-place modifier versions of string subsetting:

x %sget =% ss is the same as x <- x %sget% ss

x %strim =% ss is the same as x <- x %strim% ss

See also the documentation on string subsetting (string subset).
Note that there is no in-place modifier versions of %ss%.

Some of the internal code of these operators was inspired by the roperators R package.

Usage

x %sget =% ss

x %strim =% ss

Arguments

x, ss

see string subset.

Value

These operators do not return any value: they are in-place modifiers, and thus modify x directly.

References

Wiseman B, Nydick S, Jones J (2022). roperators: Additional Operators to Help you Write Cleaner R Code. https://CRAN.R-project.org/package=roperators

See also

Examples

ss <- c(2,2)

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
x %sget =% ss # same as x <- x %sget% ss
print(x)
#> [1] "ablm" "noyz"

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
x %strim =% ss # same as x <- x %strim% ss
print(x)
#> [1] "cdefghijk" "pqrstuvwx"


#############################################################################


ss <- c(2,2)

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
x %sget =% ss # same as x <- x %sget% ss
print(x)
#> [1] "ablm" "noyz"

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
x %strim =% ss # same as x <- x %strim% ss
print(x)
#> [1] "cdefghijk" "pqrstuvwx"