Report infix operators

The report_inops() function returns a data.frame listing the infix operators defined in the current environment (like the global environment, or the environment within a function), or a user specified environment. It also reports from which packages the infix operators came from.

report_inops(env)

Arguments

env

an optional environment to give, where the function should look for infix operators.
When not specified, the current environment (like the global environment, or the environment within a function) is used.

Value

A data.frame.

See also

Examples


if (FALSE) {
report_inops()
}