Pattern specifications for string related infix operators.

The %s-% and %s/% operators, their in-place equivalents, as well as the string detection operators (%s{}%, %s!{}%), all perform pattern matching for some purpose, where the pattern is given on the right hand side.
When a character vector or string is given on the right hand side, this is interpreted as case-sensitive regex patterns from stringi.

Instead of giving a string or character vector of regex patterns, one can also supply a list to specify exactly how the pattern should be interpreted. The list should use the exact same argument convention as stringi.

For example:

list(regex=p, case_insensitive=FALSE, ...)

list(fixed=p, ...)

list(coll=p, ...)

list(charclass=p, ...)

All arguments in the list are simply passed to the appropriate functions in stringi.
For example:
x %s/% p
counts how often regular expression specified in character vector p occurs in x, whereas the following,
x %s/% list(fixed=p, case_insensitive=TRUE)
will do the same, except it uses fixed (i.e. literal) expression, and it does not distinguish between upper case and lower case characters.

Related to the above, tinyoperations adds some convenience functions based on the stri_opts_ - functions in stringi (convenient in the sense they already have argument names, thus allowing for auto code completion):

stri_rgx(p, ...) is equivalent to c(list(regex = p), ...)

stri_fxd(p, ...) is equivalent to c(list(fixed = p), ...)

stri_cll(p, ...) is equivalent to c(list(coll = p), ...)

stri_chrcls(p, ...) is equivalent to c(list(charclass = p), ... )

With the ellipsis (...) being passed to the appropriate stri_opts-functions when it matches their arguments.

stri_rgx(
  p,
  case_insensitive,
  comments,
  dotall,
  multiline,
  time_limit,
  stack_limit,
  ...
)

stri_fxd(p, case_insensitive, overlap, ...)

stri_cll(
  p,
  locale,
  strength,
  alternate_shifted,
  french,
  uppercase_first,
  case_level,
  numeric,
  normalization,
  ...
)

stri_chrcls(p, ...)

Arguments

p

a character vector giving the pattern to search for.

case_insensitive

see stri_opts_regex and stri_opts_fixed.

comments, dotall, multiline, time_limit, stack_limit

see stri_opts_regex.

...

additional arguments not part of the stri_opts - functions to be passed here.
For example: max_count

overlap

see stri_opts_fixed.

locale, strength, alternate_shifted, french, uppercase_first, case_level, normalization, numeric

see stri_opts_collator.

Value

A list.

See also

Examples

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
p <- rep("a|e|i|o|u", 2) # same as p <- list(regex=rep("a|e|i|o|u", 2))
x %s/% p # count how often vowels appear in each string of vector x.
#> [1] 3 2

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
x %s/% list(regex = rep("A|E|I|O|U", 2), case_insensitive = TRUE)
#> [1] 3 2
x %s/% stri_rgx(rep("A|E|I|O|U", 2), case_insensitive = TRUE)
#> [1] 3 2


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
p <- list(regex = c("A", "A"), case_insensitive=TRUE)
x %s{}% p
#> [1]  TRUE FALSE
x %s!{}% p
#> [1] FALSE  TRUE

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
p <- list(fixed = c("A", "A"), case_insensitive=TRUE)
x %s{}% p
#> [1]  TRUE FALSE
x %s!{}% p
#> [1] FALSE  TRUE