String arithmetic

String arithmetic operators.

The x %s+% y operator is exported from stringi.

The x %s-% p operator removes character/pattern defined in p from x.

The x %s*% n operator is exported from stringi.

The x %s/% p operator counts how often regular expression or character pattern p occurs in each element of x.

The e1 %s$% e2 operator is exported from stringi.

x %s-% p

x %s/% p

Arguments

x

a string or character vector.

p

either a list with stringi arguments (see stri_rgx), or else a character vector of the same length as x with regular expressions.

Value

The %s+%, %s-%, and %s*% operators return a character vector of the same length as x.

The %s/% returns a integer vector of the same length as x.

The %s$% operator returns a character vector.

Details

Be aware of the precedence here!
These are not regular arithmetic; these are functions. Functions come before all arithmetic in R.
For example, the following code:
"a" %s*% 3^2
is interpreted as: ("a" %s*% 3)^2
which of course gives an error, since you cannot square a character.
Therefore, put brackets around the right hand side expression when using chaining arithmetic, like so:
"a" %s*% (3^2)

See also

Examples

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
y <- c("a", "b")
p <- rep("a|e|i|o|u", 2) # same as p <- list(regex=rep("a|e|i|o|u", 2))
n <- c(3, 2)

x %s+% y # =paste0(x,y)
#> [1] "abcdefghijklma" "nopqrstuvwxyzb"
x %s-% p # remove all vowels from x
#> [1] "bcdfghjklm"  "npqrstvwxyz"
x %s*% n
#> [1] "abcdefghijklmabcdefghijklmabcdefghijklm"
#> [2] "nopqrstuvwxyznopqrstuvwxyz"             
x %s/% p # count how often vowels appear in each string of vector x.
#> [1] 3 2


#############################################################################


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
y <- "a"
# pattern that ignores case:
p <- list(regex=rep("A|E|I|O|U", 2), case_insensitive = TRUE)
n <- c(2, 3)

x %s+% y # =paste0(x,y)
#> [1] "abcdefghijklma" "nopqrstuvwxyza"
x %s-% p # remove all vowels from x
#> [1] "bcdfghjklm"  "npqrstvwxyz"
x %s*% n
#> [1] "abcdefghijklmabcdefghijklm"             
#> [2] "nopqrstuvwxyznopqrstuvwxyznopqrstuvwxyz"
x %s/% p # count how often vowels appears in each string of vector x.
#> [1] 3 2