Concatenate Character Matrix Row-wise or Column-wise

The stri_join_mat() function (and their aliases stri_c_mat and stri_paste_mat) perform row-wise (margin=1; the default) or column-wise (margin=2) joining of a matrix of strings, thereby transforming a matrix of strings into a vector of strings.

stri_join_mat(mat, margin = 1, sep = "", collapse = NULL)

stri_c_mat(mat, margin = 1, sep = "", collapse = NULL)

stri_paste_mat(mat, margin = 1, sep = "", collapse = NULL)

Arguments

mat

a matrix of strings

margin

the margin over which the strings must be joined.

If margin=1, the elements in each row of matrix mat are joined into a single string. Thus if the matrix has 10 rows, it returns a vector of 10 strings.

If margin=2, the elements in each column of matrix mat are joined into a single string. Thus if the matrix has 10 columns, it returns a vector of 10 strings.

sep, collapse

as in stri_join.

Value

The stri_join_mat() function, and its aliases, return a vector of strings.

Details

The examples section show the uses of the stri_join_mat() function.

See also

Examples


# re-ordering characters in strings ====

x <- c("Hello world", "Goodbye world")
print(x)
#> [1] "Hello world"   "Goodbye world"
mat <- strcut_brk(x, "chr")
rank <- stringi::stri_rank(as.vector(mat)) |>  matrix(ncol=ncol(mat))
sorted <- mat %row~% rank
print(sorted)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#> [1,] ""   ""   " "  "d"  "e"  "H"  "l"  "l"  "l"  "o"   "o"   "r"   "w"  
#> [2,] " "  "b"  "d"  "d"  "e"  "G"  "l"  "o"  "o"  "o"   "r"   "w"   "y"  
stri_join_mat(sorted, margin=1)
#> [1] " deHllloorw"   " bddeGlooorwy"
stri_join_mat(sorted, margin=2)
#>  [1] " "  "b"  " d" "dd" "ee" "HG" "ll" "lo" "lo" "oo" "or" "rw" "wy"

#############################################################################

# re-ordering words ====

x <- c("Hello everyone", "Goodbye everyone")
print(x)
#> [1] "Hello everyone"   "Goodbye everyone"
mat <- strcut_brk(x, "word")
rank <- stringi::stri_rank(as.vector(mat)) |> matrix(ncol=ncol(mat))
sorted <- mat %row~% rank
print(sorted)
#>      [,1] [,2]       [,3]     
#> [1,] " "  "everyone" "Hello"  
#> [2,] " "  "everyone" "Goodbye"
stri_c_mat(sorted, margin=1) # <- alias for stri_join_mat
#> [1] " everyoneHello"   " everyoneGoodbye"
stri_c_mat(sorted, margin=2)
#> [1] "  "               "everyoneeveryone" "HelloGoodbye"    

#############################################################################

# re-ordering sentences ====

x <- c("Hello, who are you? Oh, really?! Cool!", "I don't care. But I really don't.")
print(x)
#> [1] "Hello, who are you? Oh, really?! Cool!"
#> [2] "I don't care. But I really don't."     
mat <- strcut_brk(x, "sentence")
rank <- stringi::stri_rank(as.vector(mat)) |> matrix(ncol=ncol(mat))
sorted <- mat %row~% rank
print(sorted)
#>      [,1]    [,2]                   [,3]            
#> [1,] "Cool!" "Hello, who are you? " "Oh, really?! " 
#> [2,] ""      "But I really don't."  "I don't care. "
stri_paste_mat(sorted, margin=1) # <- another alias for stri_join_mat
#> [1] "Cool!Hello, who are you? Oh, really?! "
#> [2] "But I really don't.I don't care. "     
stri_paste_mat(sorted, margin=2)
#> [1] "Cool!"                                  
#> [2] "Hello, who are you? But I really don't."
#> [3] "Oh, really?! I don't care. "