Locate \(i^{th}\) Pattern Occurrence

The stri_locate_ith function locates the \(i^{th}\) occurrence of a pattern in each string of some character vector.

stri_locate_ith(str, i, ..., regex, fixed, coll, charclass)

Arguments

str

a string or character vector.

i

a number, or a numeric vector of the same length as str. This gives the \(i^{th}\) instance to be replaced.
Positive numbers are counting from the left. Negative numbers are counting from the right. I.e.:

stri_locate_ith(str, i=1, ...)
gives the position (range) of the first occurrence of a pattern.

stri_locate_ith(str, i=-1, ...)
gives the position (range) of the last occurrence of a pattern.

stri_locate_ith(str, i=2, ...)
gives the position (range) of the second occurrence of a pattern.

stri_locate_ith(str, i=-2, ...)
gives the position (range) of the second-last occurrence of a pattern.

If abs(i) is larger than the number of instances, the first (if i < 0) or last (if i > 0) instance will be given.
For example: suppose a string has 3 instances of some pattern;
then if i >= 3 the third instance will be located,
and if i <= -3 the first instance will be located.

...

more arguments to be supplied to stri_locate and stri_count.
Do not supply the arguments omit_no_match, get_length, or pattern, as they are already specified internally. Supplying these arguments anyway will result in an error.

regex, fixed, coll, charclass

a character vector of search patterns, as in stri_locate.

Value

The stri_locate_ith() function returns an integer matrix with two columns, giving the start and end positions of the \(i^{th}\) matches, two NAs if no matches are found, and also two NAs if str is NA.

See also

Examples


# simple pattern ====

x <- rep(paste0(1:10, collapse=""), 10)
print(x)
#>  [1] "12345678910" "12345678910" "12345678910" "12345678910" "12345678910"
#>  [6] "12345678910" "12345678910" "12345678910" "12345678910" "12345678910"
out <- stri_locate_ith(x, 1:10, regex = as.character(1:10))
cbind(1:10, out)
#>          start end
#>  [1,]  1     1   1
#>  [2,]  2     2   2
#>  [3,]  3     3   3
#>  [4,]  4     4   4
#>  [5,]  5     5   5
#>  [6,]  6     6   6
#>  [7,]  7     7   7
#>  [8,]  8     8   8
#>  [9,]  9     9   9
#> [10,] 10    10  11


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
p <- rep("a|e|i|o|u",2)
out <- stri_locate_ith(x, c(-1, 1), regex=p)
print(out)
#>      start end
#> [1,]     9   9
#> [2,]     2   2
substr(x, out[,1], out[,2])
#> [1] "i" "o"


#############################################################################

# ignore case pattern ====


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
p <- rep("A|E|I|O|U", 2)
out <- stri_locate_ith(x, c(1, -1), regex=p, case_insensitive=TRUE)
substr(x, out[,1], out[,2])
#> [1] "a" "u"


#############################################################################

# multi-character pattern ====

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
#> [1] "abcdefghijklm" "nopqrstuvwxyz"
# multi-character pattern:
p <- rep("AB", 2)
out <- stri_locate_ith(x, c(1, -1), regex=p, case_insensitive=TRUE)
print(out)
#>      start end
#> [1,]     1   2
#> [2,]    NA  NA
substr(x, out[,1], out[,2])
#> [1] "ab" NA  



#############################################################################

# Replacement transformation using stringi ====

x <- c("hello world", "goodbye world")
loc <- stri_locate_ith(x, c(1, -1), regex="a|e|i|o|u")
extr <- stringi::stri_sub(x, from=loc)
repl <- chartr(extr, old = "a-zA-Z", new = "A-Za-z")
stringi::stri_sub_replace(x, loc, replacement=repl)
#> [1] "hEllo world"   "goodbye wOrld"