The tinyoperations help page

Welcome to the tinyoperations introduction help page!

The tinyoperations R-package adds adds some functions and infix operators to help in your programming etiquette.
It primarily focuses on 4 things:

(1) Safer decimal numbers ("double") truth testing (see %d==%).
(2) Extending the string manipulation capabilities of the stringi R package, see tinyoperations_strings.
(3) Reducing repetitive code; see tinyoperations_dry.
(4) A new package and module import system, that combines the benefits of aliasing a package with the benefits of attaching a package, see tinyoperations_import

And some miscellaneous functionality; see tinyoperations_misc.

The tinyoperations R-package has only one dependency, namely stringi. Most functions in this R-package are fully vectorized and optimized.

Please also have a look at the Read-Me file on the GitHub main page before using this package:
https://github.com/tony-aw/tinyoperations

tinyoperations_help()