The tinyoperations import system

The tinyoperations R package introduces a new package import system.

One can use a package without attaching the package (i.e. using :: or using a package alias), or one can attach a package (i.e. using library() or require()). The advantages and disadvantages of loading without attaching a package versus attaching a package - at least those relevant here - can be compactly presented in the following table:

aspect	|	alias / ::	|	attaching

prevent masking functions from other packages	|	Yes (+)	|	No (-)

.

prevent masking core R functions	|	Yes (+)	|	No (-)

.

clarify which function came from which package	|	Yes (+)	|	No (-)

.

place/expose functions only in current environment instead of globally	|	Yes (+)	|	No (-)

.

prevent namespace pollution	|	Yes (+)	|	No (-)

.

minimize typing - especially for infix operators

(i.e. typing package::%op%(x, y) instead of x %op% y is cumbersome)	|	No (-)	|	Yes (+)

.

use multiple related packages,

without constantly switching between package prefixes	|	No (-)	|	Yes (+)

.

NOTE: + = advantage, - = disavantage

What tinyoperations attempts to do with its import system, is to somewhat find the best of both worlds. It does this by introducing the following functions:

import_as: Allow a main package + its re-exports + its dependencies + its enhances + its extensions to be loaded under a single alias. This essentially combines the attaching advantage of using multiple related packages, whilst keeping most advantages of aliasing a package.

import_inops: Expose infix operators from a package or an alias object to the current environment. This gains the attaching advantage of less typing, whilst simultaneously avoiding the disadvantage of attaching functions from a package globally.

import_data: Directly return a data set from a package, to allow straight-forward assignment.

source_selection: Source a script, but only place the specified objects in the current environment.

There are also some additional helper functions for the package import system, see x.import and pkgs.

Usage

tinyoperations_import()

See also

Examples

if (FALSE) {
# loading "tidytable" + "data.table" under alias "tdt.":
import_as(
  tdt., "tidytable", dependencies = "data.table"
)

# exposing operators from "magrrittr" to current environment:
import_inops("magrittr")

# directly assigning the "starwars" dataset to object "d":
d <- import_data("starwars", "dplyr")

# see it in action:
d %>% tdt.$filter(species == "Droid") %>%
  tdt.$select(name, tdt.$ends_with("color"))

# source only specific functions from a script:
source_selection(list(file="sourcetest.R"), select = "helloworld")
helloworld()
}