Load main package + its re-exports + its dependencies + its enhances + its extensions under one alias

The import_as() function imports the namespace of an R package, and optionally also its re-exports, dependencies, enhances, and extensions, all under the same alias. The specified alias will be placed in the current environment (like the global environment, or the environment within a function).

Usage

import_as(
  alias,
  main_package,
  re_exports = TRUE,
  dependencies = NULL,
  enhances = NULL,
  extensions = NULL,
  lib.loc = .libPaths(),
  loadorder = c("dependencies", "main_package", "enhances", "extensions")
)

Arguments

alias

a syntactically valid non-hidden object name (unquoted), giving the alias object where the package(s) are to be loaded into.
NOTE: To keep aliases easily distinguishable from other objects that can also be subset with the $ operator, I recommend ending (not starting!) all alias names with a dot (.) or underscore (_).

main_package

a single string, giving the name of the main package to load under the given alias.

re_exports

logical; Some R packages export functions that are not defined in their own package, but in their direct dependencies; "re-exports", if you will.
This argument determines what the import_as function will do with the re-exports of the main_package:

If TRUE the re-exports from the main_package are added to the alias, even if dependencies = NULL. This is the default, as it is analogous to the behaviour of base R's :: operator.

If FALSE, these re-exports are not added, and the user must specify the appropriate packages in argument dependencies.

dependencies

an optional character vector, giving the names of the dependencies of the main_package to be loaded also under the alias.
Defaults to NULL, which means no dependencies are loaded.
See pkg_get_deps to quickly get dependencies from a package.

enhances

an optional character vector, giving the names of the packages enhanced by the main_package to be loaded also under the alias.
Defaults to NULL, which means no enhances are loaded.

extensions

an optional character vector, giving the names of the extensions of the main_package to be loaded also under the alias.
Defaults to NULL, which means no extensions are loaded.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
The lib.loc argument would usually be .libPaths().
See also loadNamespace.

loadorder

the character vector
c("dependencies", "main_package", "enhances", "extensions"),
or some re-ordering of this character vector, giving the relative load order of the groups of packages.
The default setting (which is highly recommended) is the character vector
c("dependencies", "main_package", "enhances", "extensions"),
which results in the following load order:

The dependencies, in the order specified by the depenencies argument.

The main_package (see argument main_package), including re-exports (if re_exports=TRUE).

The enhances, in the order specified by the enhances argument.

The extensions, in the order specified by the extensions argument.

Value

A locked environment object, similar to the output of loadNamespace, with the name as specified in the alias argument, will be created in the current environment (like the global environment, or the environment within a function).

To use, for example, function "some_function()" from alias "alias.", use:

alias.$some_function()

To see the special attributes of this alias object, use attr.import.

Details

On the dependencies, enhances and extensions arguments

dependencies: "Dependencies" here are defined as any package appearing in the "Depends", "Imports", or "LinkingTo" fields of the Description file of the main_package. So no recursive dependencies.

enhances: Enhances are defined as packages appearing in the "Enhances" field of the Description file of the main_package.

extensions: "Extensions" here are defined as reverse-depends or reverse-imports. It does not matter if these are CRAN or non-CRAN packages. However, the intended meaning of an extension is not merely being a reverse dependency, but a package that actually extends the functionality of the main_package.

As implied in the description of the loadorder argument, the order of the character vectors given in the dependencies, enhances, and extensions arguments matter:
If multiple packages share objects with the same name, the objects of the package named last will overwrite those of the earlier named packages.

Additional details
The import_as() function does not allow importing base/core R under an alias, so don't try.

See also

Examples

if (FALSE) {
import_as( # this creates the 'tdt.' object
  tdt., "tidytable", dependencies = "data.table"
)
tdt.$mutate
}