Expose infix operators to the current environment

The import_inops() function can do one of two things, depending on whether the pkgs argument is specified, or the delete argument is specified.

When using the import_inops() function with the pkgs argument specified, it will place the infix operators of the specified packages in the current environment (like the global environment, or the environment within a function).
(If you wish to globally attach infix operators, instead of just placing them in the current environment, see pkg_lsf.)

When using the import_inops() function with the delete argument specified, it will delete the locked infix operators with namespaces corresponding to the specified packages.
As this function specifically deletes locked infix operators, it will attempt to only remove infix operators actually exposed by import_inops() itself, thus keeping any user specified locked infix operators untouched; for more information on this, see import_lock.
Only infix operators in the current environment (like the global environment, or the environment within a function) will be deleted.

Usage

import_inops(
  pkgs,
  lib.loc = .libPaths(),
  exclude,
  include.only,
  overwrite = TRUE,
  inherits = FALSE,
  delete
)

Arguments

pkgs

a character vector of package name(s), specifying the packages from which to load infix operators, and place them in the current environment.
NOTE: The order of the character vector matters! If multiple packages share infix operators with the same name, the conflicting operators of the package named last will overwrite those of the earlier named packages.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
This is usually .libPaths().
See also loadNamespace.

exclude

a character vector, giving the infix operators NOT to expose to the current environment.
This can be handy to prevent overwriting any (user defined) infix operators already present in the current environment.

include.only

a character vector, giving the infix operators to expose to the current environment, and the rest of the operators will not be exposed.
This can be handy to prevent overwriting any (user defined) infix operators already present in the current environment.

overwrite

logical, indicating if it is allowed to overwrite existing infix operators.

If TRUE (default), a warning is given when operators existing in the current environment are being overwritten, but the function continuous nonetheless.

If FALSE, an error is produced when the to be exposed operators already exist in the current environment, and the function is halted.

inherits

logical; when overwrite=FALSE, should enclosed environments, especially package namespaces, also be taken into account?
Defaults to FALSE.
See also exists.

delete

a character vector of package name(s), specifying the infix operators from which package namespaces to delete from the current environment.
The import_inops() function will attempt not to deleted infix operators that were manually defined by the user themselves.

Value

If pkgs is specified:

The infix operators from the specified packages will be placed in the current environment (like the Global environment, or the environment within a function).

The infix operators will be LOCKED. For information on the binding lock used, see import_lock.

If delete is specified:

The infix operators from the packages specified in delete, exposed by import_inops(), will be deleted.

If such infix operators could not be found, this function returns NULL.

Details

On the exclude, include.only, overwrite, inherits arguments:
The exclude, include.only, overwrite, inherits arguments are not used when the user specifies the delete argument; they are only used when the user specifies the pkgs argument.
You cannot specify both the exclude and include.only arguments. Only one or the other, or neither.

On the pkgs and delete arguments:
You cannot specify both the pkgs and delete arguments. Only one or the other.
Unlike the import_as function, the import_inops() function does not require the packages to be necessarily related to each other.

Other details:
The import_inops() function does not support overloading base/core R operators, so don't try.

See also

Examples

if (FALSE) {
import_inops("data.table") # expose infix operators
import_inops(delete="data.table") # remove the exposed infix operators.
}