Expose (or delete/detect) infix operators to (or from/in) the current environment

The import_inops() function can do one of three things, depending on what the user specifies in argument action.

With action="expose":
When using the import_inops() function with action="expose", it will place the infix operators of the specified packages in the current environment (like the global environment, or the environment within a function).
(If you wish to globally attach infix operators, instead of just placing them in the current environment, see pkg_lsf.)

With action="remove":
When using the import_inops() function with action="remove", it will delete the infix operators in the current environment, created by import_inops(), with namespaces corresponding to the specified packages.
User-defined infix operators are not touched (attributes are used to check if an operator was created by import_inops() or not).

Usage

import_inops(pkgs, action = "expose", lib.loc = .libPaths(), ...)

Arguments

pkgs

a character vector of package name(s), specifying the packages from which to load infix operators, and place them in the current environment.
NOTE: When action="expose", the order of the character vector matters! If multiple packages share infix operators with the same name, the conflicting operators of the package named last will overwrite those of the earlier named packages.

action

a string specifying the action to perform.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
This is usually .libPaths().
See also loadNamespace.

...

additional arguments, only relevant if action = "expose".
See import_inops.control.

Value

If action = "expose":

The infix operators from the specified packages will be placed in the current environment (like the Global environment, or the environment within a function).

If action = "remove":

The infix operators from the packages specified in pkgs, exposed by import_inops(), will be deleted.

If such infix operators could not be found, this function returns NULL.

Details

Unlike the import_as function, the import_inops() function does not require the packages to be necessarily related to each other.

The import_inops() function does not support overloading base/core R operators, so don't try.

When using import_inops() to remove infix operators from the current environment, it will use the attributes of those operators to determine if the infix operator came from import_inops(), or if they were user-defined.

See also

Examples

if (FALSE) {
import_inops("data.table") # expose infix operators
import_inops("data.table", action = "remove") # remove the exposed infix operators.
}