Miscellaneous package functions

The pkgs %installed in% lib.loc operator checks if one or more package(s) pkgs exist(s) in library location lib.loc, without loading the package(s).
The syntax of this operator forces the user to make it syntactically explicit where to look for installed R package(s).

The pkg_get_deps() function gets the dependencies of a package from the Description file. It works on non-CRAN packages also.

The help.import() function finds the help file for exposed infix operators and functions in an alias object.

The pkg_lsf(package, ...) function gets a list of exported functions/operators from a package.
One handy use for this function is to, for example, globally attach all infix operators from a function using library, like so:
library(packagename, include.only = pkg_lsf("packagename", type="inops"))

Usage

pkgs %installed in% lib.loc

pkg_get_deps(
  package,
  lib.loc = .libPaths(),
  deps_type = c("LinkingTo", "Depends", "Imports"),
  base = FALSE,
  recom = FALSE,
  rstudioapi = FALSE
)

help.import(..., i, alias)

pkg_lsf(package, type, lib.loc = .libPaths())

Arguments

pkgs

a single string, or character vector, with the package name(s).

lib.loc

character vector specifying library search path (the location of R library trees to search through).
This is usually .libPaths().
See also loadNamespace.

package

a single string giving the package name.

deps_type

a character vector, giving the dependency types to be used.
Defaults to c("LinkingTo", "Depends", "Imports").
The order of the character vector given in deps_type affects the order of the returned character vector; see Details sections.

base

logical, indicating whether base/core R should be included (TRUE), or not included (FALSE; the default).

recom

logical, indicating whether the pre-installed "recommended" R packages should be included (TRUE), or not included (FALSE; the default). Note that only the recommended R packages actually installed in your system are taken into consideration.

rstudioapi

logical, indicating whether the rstudioapi R package should be included (TRUE), or not included (FALSE; the default).

...

further arguments to be passed to help.

i

either one of the following:

a function (use back-ticks when the function is an infix operator). Examples: myfun , `%operator%` , myalias.$some_function .

a string giving the function name or topic (i.e. "myfun", "thistopic"). If a string, argument alias must be specified also.

alias

the alias object as created by the import_as function. Only needs to be specified if argument i is a string, otherwise it is ignored.

type

The type of functions to list. Possibilities:

"inops" or "operators": Only infix operators.

"regfuns": Only regular functions (thus excluding infix operators).

"all": All functions, both regular functions and infix operators.

Value

For pkgs %installed in% lib.loc:

Returns a named logical vector, with the names giving the package names, and where the value TRUE indicates a package is installed, and the value FALSE indicates a package is not installed.

For pkg_get_deps():

A character vector of unique dependencies.

For pkg_lsf():

Returns a character vector of function and/or operator names.

For help.import():

Opens the appropriate help page.

Details

For help.import(...):
Do not use the topic / package and i / alias arguments together. It's either one set or the other.

For pkg_get_deps():
If using the pkgs_get_deps() function to fill in the dependencies argument of the import_as function, one may want to know the how character vector returned by pkgs_get_deps() is ordered.
The order is determined as follows.
For each string in argument deps_type, the package names in the corresponding field of the Description file are extracted, in the order as they appear in that field.
The order given in argument deps_type also affects the order of the returned character vector:
The default,
c("LinkingTo", "Depends", "Imports"),
means the package names are extracted from the fields in the following order:

"LinkingTo";

"Depends";

"Imports".

The unique (thus non-repeating) package names are then returned to the user.

References

https://stackoverflow.com/questions/30223957/elegantly-extract-r-package-dependencies-of-a-package-not-listed-on-cran

See also

Examples


if (FALSE) {
pkgs <- c(unlist(tools::package_dependencies("devtools")), "devtools")
pkgs %installed in% .libPaths()
pkg_lsf("devtools", "all")
import_as(m., "magrittr")
import_inops("magrittr")
help.import(i=m.$add)
help.import(i=`%>%`)
help.import(i="add", alias=m.)
}