The transform_if function

The transform_if() function is alternative form of ifelse. The transform_if() function transforms an object x, based on logical (TRUE, FALSE, NA) condition cond(x), such that:

For every value where cond(x)==TRUE, function yes(x) is run.

For every value where cond(x)==FALSE, function no(x) is run.

For every value where cond(x)==NA, function other(x) is run.

Usage

transform_if(
  x,
  cond,
  yes = function(x) x,
  no = function(x) x,
  other = function(x) return(NA),
  text = NULL
)

Arguments

x

a vector, matrix, or array.

cond

a (possibly anonymous) function that returns a logical vector of the same length/dimensions as x.
For example: \(x)x>0.

yes

the (possibly anonymous) transformation function to use when cond(x)==TRUE.
For example: log.
If this is not specified, yes defaults to function(x)x.

no

the (possibly anonymous) transformation function to use when cond(x)==FALSE.
For example: log.
If this is not specified, no defaults to function(x)x.

other

the (possibly anonymous) transformation function to use when cond(x) returns NA.
If this is not specified, other defaults to function(x)return(NA).
Note that other(x) is run when cond(x) is NA, not necessarily when x itself is NA!

text

a single string, to be used instead of the arguments cond, yes, no, other, in the form of
"x ; cond ; yes ; no ; other", with "x" being the declared variable.
If this string is given, cond, yes, no, other will be ignored.
The single string in question should have the following properties:

It should consist of 5 pieces of text, separated by semicolons (;).

The first piece should be a single name, declaring the name of the variable.

The second, third, fourth, and fifth pieces of text should give the expressions describing the functions to use for cond, yes, no, other, respectively, in exactly that order.

Suppose the declared variable is named x. Then, each expression in text pieces 2 to 5 will be translated as:
function(x) text.
Therefore, a function like log must be written in the text as "log(x)", not just "log".

All variables used on pieces 2 to 5 that do no match the declared variable from the first piece, are taken from the environment from which transform_if() was called.

ALL 5 pieces of text are mandatory; any missing piece results in an error.

Thus the following functions (with the declared variable being x),
cond=\(x)x > y, yes=log, no=\(x)x^2, other=\(x)-1000,
can be expressed in a single string as:
"x; x > y; log(x) ; x^2 ; -1000"

Value

Similar to ifelse. However, unlike ifelse(), the transformations are evaluated as yes(x[cond(x)]) and no(x[!cond(x)]), ensuring no unnecessary warnings or errors occur.

Details

Be careful with coercion! For example the following code:
x <- c("a", "b")
transform_if(x, \(x)x=="a", as.numeric, as.logical)
returns:
[1] NA NA
due to the same character vector being given 2 incompatible classes.

Using cond, yes, no, other directly is a tiny bit faster than using the text argument (because the text must first be translated into functions).

See also

Examples

x <- c(-10:9, NA, NA)
object_with_very_long_name <- matrix(x, ncol=2)
print(object_with_very_long_name)
#>       [,1] [,2]
#>  [1,]  -10    1
#>  [2,]   -9    2
#>  [3,]   -8    3
#>  [4,]   -7    4
#>  [5,]   -6    5
#>  [6,]   -5    6
#>  [7,]   -4    7
#>  [8,]   -3    8
#>  [9,]   -2    9
#> [10,]   -1   NA
#> [11,]    0   NA
y <- 0
z <- 1000
object_with_very_long_name |> transform_if(\(x)x>y, log, \(x)x^2, \(x)-z)
#>       [,1]          [,2]
#>  [1,]  100     0.0000000
#>  [2,]   81     0.6931472
#>  [3,]   64     1.0986123
#>  [4,]   49     1.3862944
#>  [5,]   36     1.6094379
#>  [6,]   25     1.7917595
#>  [7,]   16     1.9459101
#>  [8,]    9     2.0794415
#>  [9,]    4     2.1972246
#> [10,]    1 -1000.0000000
#> [11,]    0 -1000.0000000
object_with_very_long_name |> transform_if(text = "x ; x>y ; log(x) ; x^2 ; -z")
#>       [,1]          [,2]
#>  [1,]  100     0.0000000
#>  [2,]   81     0.6931472
#>  [3,]   64     1.0986123
#>  [4,]   49     1.3862944
#>  [5,]   36     1.6094379
#>  [6,]   25     1.7917595
#>  [7,]   16     1.9459101
#>  [8,]    9     2.0794415
#>  [9,]    4     2.1972246
#> [10,]    1 -1000.0000000
#> [11,]    0 -1000.0000000