Helper functions for the tinyoperations package import system

The help.import() function finds the help file for functions in an alias object or exposed infix operators.

The attr.import() function gets one specific special attributes or all special attributes from an alias object returned by import_as.

The is.tinyimport() function checks if an alias object or an exposed function is of class tinyimport;
i.e. if it is an object produced by the import_as or import_inops function.

Usage

help.import(..., i, alias)

attr.import(alias, which = NULL)

is.tinyimport(x)

Arguments

...

further arguments to be passed to help.

i

either one of the following:

a function (use back-ticks when the function is an infix operator). Examples: myfun , `%operator%` , myalias.$some_function . If a function, the alias argument is ignored.

a string giving the function name or topic (i.e. "myfun", "thistopic"). If a string, argument alias must be specified also.

alias

the alias object as created by the import_as function.

which

The attributes to list. If NULL, all attributes will be returned.
Possibilities: "pkgs", "conflicts", "versions", "args", and "ordered_object_names".

x

an object produced by import_as or import_inops.

Value

For help.import():

Opens the appropriate help page.

For is.tinyimport():

Returns TRUE if x is produced by functions import_as or import_inops, and returns FALSE if it is not.

For attr.import(alias, which = NULL):

All special attributes of the given alias object are returned as a list.

For attr.import(alias, which = "pkgs"):

Returns a list with 3 elements:

packages_order: a character vector of package names, giving the packages in the order they were loaded in the alias object.

main_package: a string giving the name of the main package. Re-exported functions, if present, are loaded together with the main package.

re_exports.pkgs: a character vector of package names, giving the packages from which the re-exported functions in the main package were taken.

For attr.import(alias, which = "conflicts"):

The order in which packages are loaded in the alias object (see attribute pkgs$packages_order) matters: Functions from later named packages overwrite those from earlier named packages, in case of conflicts.

The "conflicts" attribute returns a data.frame showing exactly which functions overwrite functions from earlier named packages, and as such "win" the conflicts.

For attr.import(alias, which = "versions"):

A data.frame, giving the version of every package loaded in the alias, ignoring re-exports.

For attr.import(alias, which = "args"):

Returns a list of input arguments. These were the arguments supplied to import_as when the alias object in question was created.

For attr.import(alias, which = "ordered_object_names"):

Gives the names of the objects in the alias, in the order as they were loaded.

Only unique names are given, thus conflicting objects only appear once.

(Note that if argument re_exports is TRUE, re-exported functions are loaded when the main package is loaded, thus changing this order slightly.)

Details

For help.import(...):
Do not use the topic / package and i / alias arguments together. It's either one set or the other.

See also

Examples


if (FALSE) {
import_as(mr., "magrittr")
import_inops("magrittr")

attr.import(mr.)
attr.import(mr., which="conflicts")

help.import(i=mr.$add)
help.import(i=`%>%`)
help.import(i="add", alias=mr.)
}